+++ client/src/app/users/add-user/add-user.component.ts
@@ -2,8 +2,8 @@ import { Component } from '@angular/core';
 import { FormControl, FormGroup, Validators } from '@angular/forms';
 import { MatSnackBar } from '@angular/material/snack-bar';
 import { Router } from '@angular/router';
-import { UserRole } from './user';
-import { UserService } from './user.service';
+import { UserRole } from '../user';
+import { UserService } from '../user.service';
 
 @Component({
   selector: 'app-add-user',
